refactor: moved to using src with subfolders for sdk, cli and api
