+++ src/cli/bootstrap/bootstrap.ts
@@ -1,11 +1,11 @@
 import deployContract from './deploy-contract';
 import path from 'path';
 import { createAccount, createApp } from './create-account';
-import { EosioTokenContract, setSettings } from '../src/index';
+import { EosioTokenContract, setSettings } from '../../sdk/index';
 import { signer, publicKey } from './keys';
 import bootstrapSettings from './settings';
-import settings from '../test-integration/services/settings';
-import { createUser } from '../test-integration/util/user';
+import settings from '../../../test-integration/services/settings';
+import { createUser } from '../../../test-integration/util/user';
 
 setSettings(settings);
 const eosioTokenContract = EosioTokenContract.Instance;
