+++ src/cli/bootstrap/deploy-contract.ts
@@ -3,8 +3,8 @@
 import fs from 'fs';
 import path from 'path';
 import { Name } from '@greymass/eosio';
-import { EosioContract } from '../src/index';
-import { Signer } from '../src/services/eosio/transaction';
+import { EosioContract } from '../../sdk/index';
+import { Signer } from '../../sdk/services/eosio/transaction';
 
 const eosioContract: EosioContract = EosioContract.Instance;
 
@@ -13,6 +13,7 @@ function getDeployableFilesFromDir(dir: string) {
 
     const wasmFileName = dirCont.find((filePath) => filePath.match(/.*\.(wasm)$/gi));
     const abiFileName = dirCont.find((filePath) => filePath.match(/.*\.(abi)$/gi));
+
     if (!wasmFileName) throw new Error(`Cannot find a ".wasm file" in ${dir}`);
     if (!abiFileName) throw new Error(`Cannot find an ".abi file" in ${dir}`);
 
@@ -22,11 +23,12 @@ function getDeployableFilesFromDir(dir: string) {
     };
 }
 
-async function deployContract({ account, contractDir }: { account: string, contractDir: string }, signer: Signer) {
+async function deployContract({ account, contractDir }: { account: string; contractDir: string }, signer: Signer) {
     const { wasmPath, abiPath } = getDeployableFilesFromDir(contractDir);
 
     const wasmFile = fs.readFileSync(wasmPath);
     const abiFile = fs.readFileSync(abiPath, 'utf8');
+
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     await eosioContract.deployContract(Name.from(account) as any, wasmFile, abiFile, signer);
 }
