+++ src/sdk/communication.ts
