+++ src/sdk/managers/browserStorage.ts
