+++ src/sdk/services/eosio/authority.ts
