+++ test-integration/user.test.ts
@@ -8,8 +8,8 @@ import {
     createUserObject,
     setSettings,
     EosioUtil,
-} from '../src/index';
-import { SdkErrors } from '../src/index';
+} from '../src/sdk/index';
+import { SdkErrors } from '../src/sdk/index';
 import { JsKeyManager } from '../test/services/jskeymanager';
 import { jsStorageFactory } from '../test/services/jsstorage';
 import settings from './services/settings';
