+++ test/util/crypto.test.ts
@@ -1,4 +1,4 @@
-import { sha256, randomString } from '../../src/util/crypto';
+import { sha256, randomString } from '../../src/sdk/util/crypto';
 
 describe('crypto sha256()', () => {
     it('sha256 hash', () => {
