+++ test/util/did-jwk.test.ts
@@ -1,5 +1,5 @@
 import { decodeJWT, verifyJWT } from '@tonomy/did-jwt';
-import { resolve } from '../../src/util/did-jwk';
+import { resolve } from '../../src/sdk/util/did-jwk';
 
 describe('did-jwk resolver', () => {
     const jwt =
