+++ test/util/message.test.ts
@@ -1,7 +1,7 @@
 import { PrivateKey } from '@greymass/eosio';
 import { createSigner } from '@tonomy/antelope-ssi-toolkit';
-import { createVCSigner, generateRandomKeyPair, KeyManagerLevel, setSettings } from '../../src';
-import { Message } from '../../src/util/message';
+import { createVCSigner, generateRandomKeyPair, KeyManagerLevel, setSettings } from '../../src/sdk';
+import { Message } from '../../src/sdk/util/message';
 import { JsKeyManager } from '../services/jskeymanager';
 
 setSettings({
